requirements.txt
anthropic
python-dotenv
schedule
streamlit
secure-smtplib